requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-autocomplete-light==3.8.2
django-bootstrap3==15.0.0
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-widget-tweaks==1.4.8
idna==3.3
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
xlwt==1.3.0
